ALS Gradient Accumulation Support and Tests.
Summary:
A couple of extra test cases for GA with ALS.

ALS Gradient Accumulation Tests.

A couple of extra test cases for GA with ALS and a wrapper to facilitate.

Test Plan: CI

Reviewers: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Reviewed By: #tensorflow, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, alfiee

Maniphest Tasks: T64439

Differential Revision: https://phabricator.sourcevertex.net/D73024

+++ keras/ipu/layers/capture_upstream_gradients.py
@@ -35,7 +35,7 @@ class _GradientCaptureLayerBase(Layer):
     self._layer = layer
     self._tag = tag
 
-    super().__init__(name=tag)
+    super().__init__(dtype=layer.dtype, name=tag)
 
   def call(self, inputs, *args, **kwargs):
     raise NotImplementedError(
